+++ packages/core/src/translation/ar.ts
@@ -1,0 +1,20 @@
+import type { TranslationInterface } from '.';
+
+export const arLanguage: TranslationInterface = {
+  write_a_message_placeholder: 'اكتب رسالة...',
+  your_issue_has_been_resolved: 'تم حل مشكلتك!',
+  new_conversation: 'محادثة جديدة',
+  welcome_screen_title: 'مرحبًا بك في دردشة الدعم الخاصة بنا',
+  welcome_screen_description:
+    'نحن هنا للمساعدة! ابدأ محادثة وسنرد عليك في أقرب وقت ممكن.',
+  your_name_placeholder: 'اسمك',
+  your_email_placeholder: 'عنوان بريدك الإلكتروني',
+  start_chat_button: 'تحدث إلى الدعم',
+  start_chat_button_loading: 'جاري الاتصال...',
+  i_need_more_help: 'أحتاج المزيد من المساعدة',
+  this_was_helpful: 'كان هذا مفيدًا',
+  optional: 'اختياري',
+  no_conversations_yet: 'لا يوجد محادثات',
+  back_to_conversations: 'العودة إلى المحادثات',
+  closed_conversations: 'المحادثات المغلقة',
+};