+++ packages/core/src/translation/de.ts
@@ -1,0 +1,20 @@
+import type { TranslationInterface } from '.';
+
+export const deLanguage: TranslationInterface = {
+  write_a_message_placeholder: 'Nachricht schreiben...',
+  your_issue_has_been_resolved: 'Ihr Problem wurde gelöst!',
+  new_conversation: 'Neue Konversation',
+  welcome_screen_title: 'Willkommen in unserem Support-Chat',
+  welcome_screen_description:
+    'Wir sind hier, um zu helfen! Beginnen Sie ein Gesprách und wir werden so schnell wie mogelijk antworten.',
+  your_name_placeholder: 'Ihr Name',
+  your_email_placeholder: 'Ihre E-Mail-Adresse',
+  start_chat_button: 'Mit dem Support sprechen',
+  start_chat_button_loading: 'Verbindung wird hergestellt...',
+  i_need_more_help: 'Ich brauche weitere Hilfe',
+  this_was_helpful: 'Dies war hilfreich',
+  optional: 'Optional',
+  no_conversations_yet: 'noch keine Gespräche',
+  back_to_conversations: 'Zurück zur Konversationen',
+  closed_conversations: 'Geschlossene Konversationen',
+};