+++ packages/core/src/translation/en.ts
@@ -1,0 +1,20 @@
+import type { TranslationInterface } from '.';
+
+export const enLanguage: TranslationInterface = {
+  write_a_message_placeholder: 'Write a message...',
+  your_issue_has_been_resolved: 'Your issue has been resolved!',
+  new_conversation: 'New conversation',
+  welcome_screen_title: 'Welcome to our support chat',
+  welcome_screen_description:
+    "We're here to help! Start a conversation and we'll get back to you as soon as possible.",
+  your_name_placeholder: 'Your name',
+  your_email_placeholder: 'Your email address',
+  start_chat_button: 'Talk to support',
+  start_chat_button_loading: 'Connecting...',
+  i_need_more_help: 'I need more help',
+  this_was_helpful: 'This was helpful',
+  optional: 'Optional',
+  no_conversations_yet: 'No conversations yet',
+  back_to_conversations: 'Back to conversations',
+  closed_conversations: 'Closed conversations',
+};