+++ packages/core/src/translation/fr.ts
@@ -1,0 +1,20 @@
+import type { TranslationInterface } from '.';
+
+export const frLanguage: TranslationInterface = {
+  write_a_message_placeholder: 'Écrivez un message...',
+  your_issue_has_been_resolved: 'Votre problème a été résolu !',
+  new_conversation: 'Nouvelle conversation',
+  welcome_screen_title: 'Bienvenue dans notre chat de support',
+  welcome_screen_description:
+    'Nous sommes là pour vous aider ! Commencez une conversation et nous vous répondrons dès que possible.',
+  your_name_placeholder: 'Votre nom',
+  your_email_placeholder: 'Votre adresse e-mail',
+  start_chat_button: 'Parler au support',
+  start_chat_button_loading: 'Connexion...',
+  i_need_more_help: "Je besoin d'aide plus",
+  this_was_helpful: "C'était utile",
+  optional: 'Optionnel',
+  no_conversations_yet: 'Aucune conversation pour le moment',
+  back_to_conversations: 'Retour aux conversations',
+  closed_conversations: 'Conversations fermées',
+};