+++ packages/core/src/translation/nl.ts
@@ -1,0 +1,20 @@
+import type { TranslationInterface } from '.';
+
+export const nlLanguage: TranslationInterface = {
+  write_a_message_placeholder: 'Schrijf een bericht...',
+  your_issue_has_been_resolved: 'Uw probleem is opgelost!',
+  new_conversation: 'Nieuw gesprek',
+  welcome_screen_title: 'Welkom bij onze supportchat',
+  welcome_screen_description:
+    'We zijn hier om te helpen! Begin een gesprek en we nemen zo snel mogelijk contact met u op.',
+  your_name_placeholder: 'Uw naam',
+  your_email_placeholder: 'Uw e-mailadres',
+  start_chat_button: 'Praat met ondersteuning',
+  start_chat_button_loading: 'Verbinding maken...',
+  i_need_more_help: 'Ik heb nog meer hulp nodig',
+  this_was_helpful: 'Mijn vraag is opgelost',
+  optional: 'Optioneel',
+  no_conversations_yet: 'Nog geen gesprekken',
+  back_to_conversations: 'Terug naar gesprekken',
+  closed_conversations: 'Afgesloten gesprekken',
+};